Refactor: split public.py into modular route files
- Split public.py (1124 lines) into 6 focused modules:
  - dashboard.py: Home page and personalized dashboard
  - schedule_personal.py: Individual schedule views (day/week/month/year)
  - schedule_all.py: Team-wide schedule views
  - schedule_api.py: API endpoints for lazy loading
  - overtime.py: Overtime shift management
  - shared.py: Shared templates and Jinja utilities
- Update main.py router imports to use new modular structure
- Improve code organization and maintainability
