+++ app/main.py
@@ -21,7 +21,11 @@ from app.core.sentry_config import init_sentry
 from app.database.database import create_tables, get_db
 from app.routes.admin import router as admin_router
 from app.routes.auth_routes import router as auth_router
-from app.routes.public import router as public_router
+from app.routes.dashboard import router as dashboard_router
+from app.routes.overtime import router as overtime_router
+from app.routes.schedule_all import router as schedule_all_router
+from app.routes.schedule_api import router as schedule_api_router
+from app.routes.schedule_personal import router as schedule_personal_router
 
 # Setup logging FIRST (before any other imports that might log)
 setup_logging()
@@ -157,7 +161,11 @@ app.add_middleware(RequestLoggingMiddleware)
 app.mount("/static", StaticFiles(directory="app/static"), name="static")
 
 # Include routers
-app.include_router(public_router)
+app.include_router(dashboard_router)
+app.include_router(schedule_personal_router)
+app.include_router(schedule_all_router)
+app.include_router(schedule_api_router)
+app.include_router(overtime_router)
 app.include_router(auth_router)
 app.include_router(admin_router)
